feat(calendar): add calendar
There are some caveats to the calendar but it works. It can display the different events created.
The caveat is that it is currently displaying in the current timezone meaning that when creating the
todos, the dueDate field is changed to fit the current timezone so that the calendar can render the
time and match them.

closed #42